25/05 15:53

+++ app/src/main/java/com/lampro/myaccuweather/objects/currentweatherresponse/Main.kt
@@ -15,9 +15,9 @@ data class Main(
     @SerializedName("sea_level")
     var seaLevel: Int,
     @SerializedName("temp")
-    var temp: Int,
+    var temp: Double,
     @SerializedName("temp_max")
-    var tempMax: Int,
+    var tempMax: Double,
     @SerializedName("temp_min")
-    var tempMin: Int
+    var tempMin: Double
 )
